Refactor: Separate Header.jsx component - Header 컴포넌트를 jsx 파일로 분리
